pyproject.toml
[tool.poetry]
name = "companyproject"
version = "0.1.0"
description = ""
authors = []
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^4.1.7"
djangorestframework = "^3.14.0"


[tool.poetry.group.dev.dependencies]
ipdb = "^0.13.11"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
